Refactor: Error 객체 추가
eslint, prettier 적용
check 함수 파일 분리
app.js 루트폴더로 이동
