+++ .prettierrc
@@ -1,0 +1,8 @@
+{
+  "tabWidth": 2,
+  "semi": true,
+  "arrowParens": "avoid",
+  "printWidth": 120,
+  "trailingComma": "all",
+  "singleQuote": false
+}
