+++ config/routes.js
@@ -4,22 +4,23 @@ const { postsRouter } = require("../src/posts/post.route");
 const { authRouter } = require("../src/auth/auth.route");
 const { commentsRouter } = require("../src/comments/comment.route");
 const logger = require("./logger");
-const { errorHandlerWrapper } = require("../src/middlewares/errorhandler");
-const { serverCheck } = require("../src/middlewares/serverCheck");
+const { errorHandlerWrapper, errorTester } = require("../src/middlewares/errorhandler");
+const { serverCheck } = require("../src/middlewares/checks");
 
 app.use("/posts", postsRouter);
 app.use("/auth", authRouter);
 app.use("/comments", commentsRouter);
 
-app.use("/", errorHandlerWrapper(serverCheck));
+app.get("/", errorHandlerWrapper(serverCheck));
+app.get("/test", errorHandlerWrapper(errorTester));
 
 app.use("*", (req, res, next) => {
   res.send("404 Not Found");
 });
 
 app.use((error, req, res, next) => {
-  logger.error(error.message);
-  res.status(500).send({ message: "Server Error", error });
+  res.status(500).send(error);
+  logger.error(error);
 });
 
 module.exports = app;
