+++ src/auth/auth.controller.js
@@ -1,3 +1,4 @@
+const errors = require("../../errors");
 const AuthService = require("./auth.service");
 const authService = new AuthService();
 
@@ -5,7 +6,7 @@ exports.postSignup = async (req, res, next) => {
   const { email, password } = req.body;
 
   if (!email || !password) {
-    next("error: no body");
+    next(errors.notFound("email, password를 채워주세요."));
   }
 
   await authService.postSignup(email, password);
@@ -16,7 +17,7 @@ exports.postSignin = async (req, res, next) => {
   const { email, password } = req.body;
 
   if (!email || !password) {
-    next("error: no body");
+    next(errors.notFound("email, password를 채워주세요."));
   }
 
   const { accessToken, user_id } = await authService.postSignin(email, password);
