+++ src/middlewares/checks.js
@@ -9,3 +9,11 @@ exports.serverCheck = async (req, res) => {
 
   return res.json({ rows });
 };
+
+exports.userCheck = (req, res) => {
+  if (req.session.user) {
+    res.json({ isLogin: true, user: req.session.user });
+  } else {
+    res.json({ isLogin: false });
+  }
+};
